mozjs24: Update of double-conversion library
Update to the latest sources in the upstream double-conversion
library (https://github.com/google/double-conversion). This is required
to build on aarch64 hardware, otherwise you get the following build errors:

  "Target architecture was not detected as supported by Double-Conversion."

This was fixed in upstream mozilla some time after mozjs24 with a patch
on this bug report:
    https://bugzilla.mozilla.org/show_bug.cgi?id=962488

That patch does not apply against the mozjs24 tarball though, so resorted
to manually updating the double-precision directory following the script
at mozjs24/mfbt/double-conversion/update.sh.