gstreamer plugins: Tune builds for i386 and arm
GStreamer plugins tend to produce assembly code in some places, for
this reason it's important to explicitly specify the build triple
so that it chooses to generate the right code for the target platform.

Note: on i386 we specify i586-unknown-linux-gnu because that is
the triple of the compiler we have in the base runtime.