Clean up webkitgtk build definition
-DNDEBUG should already be defined since the build type is RELEASE

Messing with SHARED_LIBS is not supported; we're not going to install
any static libraries, but we do need them during the build. No clue what
happens when playing with this variable.

All the other options are already being set to their default values.

Remove the buildapi patch, seems to be unused now.