Attempt to upgrade to WebKitGTK+ 2.19.3
We finally have a successful master build for the first time in a long
time. But WebKit's memory allocator is broken in 2.19.2, which is not
very great. There's a new release out, so let's try upgrading.

Also, temporarily remove my inhibitor debugging patch. It will need to
be updated for 2.19.3.