[cantarell-fonts] Use Autotools back.
I’m not happy with disabling the useprebuilt=true option,
but it is complaining about missing (prebuilt) files, and
I don’t know enough about this module to try to fix that.

Instead, use Autotools and the last tarball release, like
in 3.26 branch. Let’s see if that helps fixing the build.