Switch vala to build from tarball
Something pushed to 0.40 branch on June 13 broke vala such that meson is
no longer able to detect a valid version of vala installed. Previously,
meson detected:

Native Vala compiler: valac (valac 0.40.6-dirty)

But now it detects:

Native Vala compiler: valac (valac unknown version)

I suspect this is not actually related to anything that was pushed, but
rather to the fact that commits were pushed at all, because vala uses
'git describe' to compute its version. The meson support for that is
probably busted somehow.

This means we will need to build vala from tarball. This is sad, as we
are not going to keep it updated and will only be testing GNOME with an
old version of vala for the foreseeable future.